+++ src/customisations/components/views/messages/ContentScanningFileBody.tsx
@@ -25,8 +25,8 @@ import { _t } from "matrix-react-sdk/src/languageHandler";
 import { presentableTextForFile } from "matrix-react-sdk/src/utils/FileUtils";
 
 import { Media } from "../../../ContentScanningMedia";
-import OriginalFileBody from "../../../../components/views/messages/OriginalFileBody";
-import { ContentScanningStatus } from "../../../../components/views/elements/ContentScanningStatus";
+import OriginalFileBody from "../../../../tchap/components/views/messages/OriginalFileBody";
+import { ContentScanningStatus } from "../../../../tchap/components/views/elements/ContentScanningStatus";
 
 interface Props extends IBodyProps {
     /* whether or not to show the default placeholder for the file. Defaults to true. */
