+++ src/customisations/components/views/messages/ContentScanningStickerBody.tsx
@@ -23,10 +23,10 @@ import { IMediaEventContent } from "matrix-react-sdk/src/customisations/models/I
 import { presentableTextForFile } from "matrix-react-sdk/src/utils/FileUtils";
 import { _t } from "matrix-react-sdk/src/languageHandler";
 
-import OriginalImageBody from "../../../../components/views/messages/OriginalImageBody";
+import OriginalImageBody from "../../../../tchap/components/views/messages/OriginalImageBody";
 import { Media } from "../../../ContentScanningMedia";
-import { BlockedIcon } from "../../../../components/views/elements/BlockedIcon";
-import { ContentScanningStatus } from "../../../../components/views/elements/ContentScanningStatus";
+import { BlockedIcon } from "../../../../tchap/components/views/elements/BlockedIcon";
+import { ContentScanningStatus } from "../../../../tchap/components/views/elements/ContentScanningStatus";
 
 interface State {
     isScanning: boolean;
