+++ src/tchap/async-components/views/dialogs/security/TchapExportE2eKeysDialog.tsx
@@ -26,7 +26,7 @@ import { KeysStartingWith } from "matrix-react-sdk/src/@types/common";
 import { MatrixClient } from "matrix-js-sdk/src/client";
 import { logger } from "matrix-js-sdk/src/logger";
 
-import "../../../../../res/css/views/dialogs/_TchapExportE2eKeysDialog.pcss";
+import "../../../../../../res/css/views/dialogs/_TchapExportE2eKeysDialog.pcss";
 
 enum Phase {
     Edit = "edit",
