+++ src/tchap/components/views/common/TchapAnchor.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import "../../../res/css/common/_TchapAnchor.pcss";
+import "../../../../../res/css/common/_TchapAnchor.pcss";
 
 interface TchapAnchorProps {
     children: React.ReactNode;
