+++ src/tchap/components/views/dialogs/TchapCreateRoomDialog.tsx
@@ -23,10 +23,10 @@ import Field from "matrix-react-sdk/src/components/views/elements/Field";
 import DialogButtons from "matrix-react-sdk/src/components/views/elements/DialogButtons";
 import BaseDialog from "matrix-react-sdk/src/components/views/dialogs/BaseDialog";
 
-import TchapUtils from "../../../tchap/util/TchapUtils";
-import TchapRoomTypeSelector from "./../elements/TchapRoomTypeSelector";
-import { TchapRoomType } from "../../../tchap/@types/tchap";
-import TchapCreateRoom from "../../../tchap/lib/createTchapRoom";
+import TchapUtils from "../../../util/TchapUtils";
+import TchapRoomTypeSelector from "../elements/TchapRoomTypeSelector";
+import { TchapRoomType } from "../../../@types/tchap";
+import TchapCreateRoom from "../../../lib/createTchapRoom";
 
 // We leave the same props as Element's version, to avoid unknown props warnings.
 interface IProps {
