+++ src/tchap/components/views/elements/TchapRoomTypeSelector.tsx
@@ -8,9 +8,9 @@ import { _t } from "matrix-react-sdk/src/languageHandler";
 import StyledRadioButton from "matrix-react-sdk/src/components/views/elements/StyledRadioButton";
 import LabelledToggleSwitch from "matrix-react-sdk/src/components/views/elements/LabelledToggleSwitch";
 
-import { TchapRoomType } from "../../../tchap/@types/tchap";
+import { TchapRoomType } from "../../../@types/tchap";
 
-import "../../../../res/css/views/elements/_TchapRoomTypeSelector.pcss";
+import "../../../../../res/css/views/elements/_TchapRoomTypeSelector.pcss";
 
 interface IProps {
     value: TchapRoomType;
