+++ src/tchap/components/views/messages/OriginalImageBody.tsx
