+++ src/tchap/components/views/messages/OriginalStickerBody.tsx
