+++ src/tchap/lib/ExpiredAccountHandler.ts
@@ -6,7 +6,7 @@ import { MatrixClientPeg } from "matrix-react-sdk/src/MatrixClientPeg";
 import Modal from "matrix-react-sdk/src/Modal";
 import PlatformPeg from "matrix-react-sdk/src/PlatformPeg";
 
-import ExpiredAccountDialog from "../../components/views/dialogs/ExpiredAccountDialog";
+import ExpiredAccountDialog from "../components/views/dialogs/ExpiredAccountDialog";
 import TchapUtils from "../util/TchapUtils";
 
 /*
