+++ test/unit-tests/components/views/settings/TchapJoinRuleSettings-test.tsx
@@ -9,7 +9,7 @@ import {
 } from "matrix-react-sdk/test/test-utils/test-utils";
 import { JoinRule, MatrixClient, Room } from "matrix-js-sdk/src/matrix";
 
-import TchapJoinRuleSettings from "../../../../../src/components/views/settings/TchapJoinRuleSettings";
+import TchapJoinRuleSettings from "../../../../../src/tchap/components/views/settings/TchapJoinRuleSettings";
 import { TchapRoomAccessRule, TchapRoomAccessRulesEventId } from "../../../../../src/tchap/@types/tchap";
 
 function mkStubRoomWithInviteRule(roomId: string, name: string, client: MatrixClient, joinRule: JoinRule): Room {
